Add files via upload
The MNIST dataset consists of collection o gray scale images representing handwritten digits from 0 to 9 . this is consists of two main components : a training set and testing set. The purpose of the MNIST dataset is to provide a standardized and representative dataset for training. The models achieved high accuracy and demonstrated the ability to classify handwritten digits effectively. Evaluation metrics such as accuracy, precision, recall, and F1 score were used to assess the models' performance. The Random Forest Classifier was used to determine the importance of features. Principal Component Analysis (PCA) was applied to visualize the data in a lower-dimensional space.Overall, the analysis of the MNIST dataset and the obtained results provide a strong foundation for digit classification tasks. By exploring different models, refining hyperparameters, and considering future directions, further advancements can be made to enhance the accuracy and applicability of the models in real-world scenarios.